Moved stats card to its own component
